[NoWork -> noWorkResult] and change implementation
